+++ src/manager/utils/doBye.js
@@ -1,0 +1,5 @@
+import { dictionary as dict } from '../../costants.js';
+
+export const doBye = (username) => {
+  console.log(dict.GET_BYE_TEXT(username));
+};
