+++ src/manager/utils/showErrorMsgOnExit.js
@@ -1,0 +1,5 @@
+import { dictionary as dict } from '../../costants.js';
+
+export const showErrorMsgOnExit = () => {
+  console.log(dict.ERROR_MSG_ON_EXIT);
+};
