+++ src/manager/utils/showErrorMsgOnOperationFailed.js
@@ -1,0 +1,5 @@
+import { dictionary as dict } from '../../costants.js';
+
+export const showErrorMsgOnOperationFailed = () => {
+  console.log(dict.ERROR_MSG_ON_OPERATION_FAILED);
+};
